FEAT: sesión 09/06/2026

+++ app/src/main/java/com/uam/taskvmg1/screen/TaskListScreen.kt
@@ -85,7 +85,7 @@ fun TaskListScreen(navController: NavController,
                             .fillMaxWidth()
                             .clickable()
                             {
-                                navController.navigate(TaskDetail(viewModel.tasks[it].id))
+                                navController.navigate(TaskDetail(viewModel.tasks[it].id.toInt()))
                             }
 
                     )
